# backend/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Grinch API
  version: 0.1.0
servers:
  - url: /api/v1
security:
  - sessionAuth: []
paths:
  /machines:
    get:
      operationId: listMachines
      tags:
        - machines
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/UserIdFilter'
      responses:
        '200':
          description: Machine list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineListResponse'
        default:
          $ref: '#/components/responses/Problem'
  /machines/{id}:
    get:
      operationId: getMachine
      tags:
        - machines
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Machine detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Machine'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteMachine
      tags:
        - machines
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Machine deleted.
        default:
          $ref: '#/components/responses/Problem'
  /users:
    get:
      operationId: listUsers
      tags:
        - users
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
      responses:
        '200':
          description: User list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        default:
          $ref: '#/components/responses/Problem'
  /users/{id}:
    get:
      operationId: getUser
      tags:
        - users
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: User detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
  /groups:
    get:
      operationId: listGroups
      tags:
        - groups
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
      responses:
        '200':
          description: Group list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createGroup
      tags:
        - groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreateRequest'
      responses:
        '201':
          description: Group created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        default:
          $ref: '#/components/responses/Problem'
  /groups/{id}:
    get:
      operationId: getGroup
      tags:
        - groups
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Group detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    patch:
      operationId: patchGroup
      tags:
        - groups
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupPatchRequest'
      responses:
        '200':
          description: Group updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteGroup
      tags:
        - groups
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Group deleted.
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
  /group-memberships:
    get:
      operationId: listGroupMemberships
      tags:
        - group-memberships
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/GroupIdFilter'
        - $ref: '#/components/parameters/UserIdFilter'
        - $ref: '#/components/parameters/MachineIdFilter'
      responses:
        '200':
          description: Group membership list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupMembershipListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createGroupMembership
      tags:
        - group-memberships
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupMembershipCreateRequest'
      responses:
        '201':
          description: Group membership created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupMembership'
        default:
          $ref: '#/components/responses/Problem'
  /group-memberships/{id}:
    get:
      operationId: getGroupMembership
      tags:
        - group-memberships
      parameters:
        - $ref: '#/components/parameters/GroupMembershipId'
      responses:
        '200':
          description: Group membership detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupMembership'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteGroupMembership
      tags:
        - group-memberships
      parameters:
        - $ref: '#/components/parameters/GroupMembershipId'
      responses:
        '204':
          description: Group membership deleted.
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
  /rules:
    get:
      operationId: listRules
      tags:
        - rules
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
      responses:
        '200':
          description: Rule list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createRule
      tags:
        - rules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleCreateRequest'
      responses:
        '201':
          description: Rule created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        default:
          $ref: '#/components/responses/Problem'
  /rules/{id}:
    get:
      operationId: getRule
      tags:
        - rules
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Rule detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    patch:
      operationId: patchRule
      tags:
        - rules
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RulePatchRequest'
      responses:
        '200':
          description: Rule updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteRule
      tags:
        - rules
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Rule deleted.
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
  /rule-targets:
    get:
      operationId: listRuleTargets
      tags:
        - rule-targets
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/RuleIdFilter'
        - $ref: '#/components/parameters/SubjectKindFilter'
        - $ref: '#/components/parameters/SubjectIdFilter'
        - $ref: '#/components/parameters/RuleTargetAssignmentFilter'
        - $ref: '#/components/parameters/RulePolicyFilter'
      responses:
        '200':
          description: Rule target list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleTargetListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createRuleTarget
      tags:
        - rule-targets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleTargetCreateRequest'
      responses:
        '201':
          description: Rule target created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleTarget'
        default:
          $ref: '#/components/responses/Problem'
  /rule-targets/{id}:
    get:
      operationId: getRuleTarget
      tags:
        - rule-targets
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Rule target detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleTarget'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    patch:
      operationId: patchRuleTarget
      tags:
        - rule-targets
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleTargetPatchRequest'
      responses:
        '200':
          description: Rule target updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleTarget'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteRuleTarget
      tags:
        - rule-targets
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Rule target deleted.
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
  /executables:
    get:
      operationId: listExecutables
      tags:
        - executables
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
      responses:
        '200':
          description: Executable list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutableListResponse'
        default:
          $ref: '#/components/responses/Problem'
  /executables/{id}:
    get:
      operationId: getExecutable
      tags:
        - executables
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Executable detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Executable'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
  /execution-events:
    get:
      operationId: listExecutionEvents
      tags:
        - execution-events
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/MachineIdFilter'
        - $ref: '#/components/parameters/UserIdFilter'
        - $ref: '#/components/parameters/ExecutableIdFilter'
      responses:
        '200':
          description: Execution event list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionEventListResponse'
        default:
          $ref: '#/components/responses/Problem'
  /execution-events/{id}:
    get:
      operationId: getExecutionEvent
      tags:
        - execution-events
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Execution event detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionEvent'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteExecutionEvent
      tags:
        - execution-events
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Execution event deleted.
        default:
          $ref: '#/components/responses/Problem'
  /file-access-events:
    get:
      operationId: listFileAccessEvents
      tags:
        - file-access-events
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/MachineIdFilter'
        - $ref: '#/components/parameters/ExecutableIdFilter'
      responses:
        '200':
          description: File access event list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileAccessEventListResponse'
        default:
          $ref: '#/components/responses/Problem'
  /file-access-events/{id}:
    get:
      operationId: getFileAccessEvent
      tags:
        - file-access-events
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: File access event detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileAccessEvent'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteFileAccessEvent
      tags:
        - file-access-events
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: File access event deleted.
        default:
          $ref: '#/components/responses/Problem'
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: grinch_session
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        format: int32
        minimum: 1
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        format: int32
        minimum: 0
    Search:
      name: search
      in: query
      schema:
        type: string
    Sort:
      name: sort
      in: query
      schema:
        type: string
      description: Sort field name.
    Order:
      name: order
      in: query
      schema:
        type: string
        enum:
          - asc
          - desc
    UserIdFilter:
      name: user_id
      in: query
      schema:
        type: string
        format: uuid
    MachineIdFilter:
      name: machine_id
      in: query
      schema:
        type: string
        format: uuid
    ExecutableIdFilter:
      name: executable_id
      in: query
      schema:
        type: string
        format: uuid
    GroupIdFilter:
      name: group_id
      in: query
      schema:
        type: string
        format: uuid
    RuleIdFilter:
      name: rule_id
      in: query
      schema:
        type: string
        format: uuid
    SubjectIdFilter:
      name: subject_id
      in: query
      schema:
        type: string
        format: uuid
    GroupMembershipId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SubjectKindFilter:
      name: subject_kind
      in: query
      schema:
        $ref: '#/components/schemas/RuleTargetSubjectKind'
    RuleTargetAssignmentFilter:
      name: assignment
      in: query
      schema:
        $ref: '#/components/schemas/RuleTargetAssignment'
    RulePolicyFilter:
      name: policy
      in: query
      schema:
        $ref: '#/components/schemas/RulePolicy'
  responses:
    Problem:
      description: Problem response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Machine:
      type: object
      required:
        - id
        - serial_number
        - hostname
        - model_identifier
        - os_version
        - os_build
        - santa_version
        - primary_user
        - request_clean_sync
        - last_seen_at
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        serial_number:
          type: string
        hostname:
          type: string
        model_identifier:
          type: string
        os_version:
          type: string
        os_build:
          type: string
        santa_version:
          type: string
        primary_user:
          type: string
        primary_user_id:
          type: string
          format: uuid
          nullable: true
        request_clean_sync:
          type: boolean
        last_seen_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    User:
      type: object
      required:
        - id
        - upn
        - display_name
        - source
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        upn:
          type: string
        display_name:
          type: string
        source:
          $ref: '#/components/schemas/Source'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Group:
      type: object
      required:
        - id
        - name
        - description
        - source
        - member_count
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        source:
          $ref: '#/components/schemas/Source'
        member_count:
          type: integer
          format: int32
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    GroupCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
    GroupPatchRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
        description:
          type: string
...
    GroupMembership:
      type: object
      required:
        - id
        - kind
        - group
        - member
        - created_at
        - updated_at
      properties:
        id:
          type: string
        kind:
          $ref: '#/components/schemas/GroupMembershipKind'
        group:
          $ref: '#/components/schemas/GroupMembershipGroup'
        member:
          $ref: '#/components/schemas/GroupMembershipMember'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    GroupMembershipCreateRequest:
      type: object
      required:
        - group_id
        - member_kind
        - member_id
      properties:
        group_id:
          type: string
          format: uuid
        member_kind:
          $ref: '#/components/schemas/MemberKind'
        member_id:
          type: string
          format: uuid
...
    Rule:
      allOf:
        - $ref: '#/components/schemas/RuleSummary'
        - type: object
          required:
            - custom_message
            - custom_url
          properties:
            custom_message:
              type: string
            custom_url:
              type: string
    RuleCreateRequest:
      type: object
      required:
        - name
        - rule_type
        - identifier
      properties:
        name:
          type: string
        description:
          type: string
        rule_type:
          $ref: '#/components/schemas/RuleType'
        identifier:
          type: string
        custom_message:
          type: string
        custom_url:
          type: string
    RulePatchRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
        description:
          type: string
        rule_type:
          $ref: '#/components/schemas/RuleType'
        identifier:
          type: string
        custom_message:
          type: string
        custom_url:
          type: string
...
    RuleTargetCreateRequest:
      type: object
      required:
        - rule_id
        - subject_id
        - assignment
      properties:
        rule_id:
          type: string
          format: uuid
        subject_id:
          type: string
          format: uuid
        assignment:
          $ref: '#/components/schemas/RuleTargetAssignment'
        priority:
          type: integer
          format: int32
        policy:
          $ref: '#/components/schemas/RulePolicy'
        cel_expression:
          type: string
...
    MachineListResponse:
      type: object
      required:
        - total
        - rows
      properties:
        total:
          type: integer
          format: int32
        rows:
          type: array
          items:
            $ref: '#/components/schemas/MachineSummary'
    UserListResponse:
      type: object
      required:
        - total
        - rows
      properties:
        total:
          type: integer
          format: int32
        rows:
          type: array
          items:
            $ref: '#/components/schemas/User'
    GroupListResponse:
      type: object
      required:
        - total
        - rows
      properties:
        total:
          type: integer
          format: int32
        rows:
          type: array
          items:
            $ref: '#/components/schemas/Group'
    GroupMembershipListResponse:
      type: object
      required:
        - total
        - rows
      properties:
        total:
          type: integer
          format: int32
        rows:
          type: array
          items:
            $ref: '#/components/schemas/GroupMembership'
    RuleListResponse:
      type: object
      required:
        - total
        - rows
      properties:
        total:
          type: integer
          format: int32
        rows:
          type: array
          items:
            $ref: '#/components/schemas/RuleSummary'
    RuleTargetListResponse:
      type: object
      required:
        - total
        - rows
      properties:
        total:
          type: integer
          format: int32
        rows:
          type: array
          items:
            $ref: '#/components/schemas/RuleTargetSummary'
...